pyproject.toml
[project]
name = "minecraft-translator-flet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flet-cli==0.28.3",
    "flet-desktop==0.28.3",
    "flet-web==0.28.3",
    "flet[all]==0.28.3",
    "ftb-snbt-lib>=0.4.1",
    "google-genai>=1.56.0",
    "httpx>=0.28.1",
    "opencc-python-reimplemented>=0.1.7",
    "orjson>=3.11.3",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pytest>=9.0.2",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "wikipedia>=1.4.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
